requirements.txt
## requirements for thankr##
django
djangorestframework
textblob